+++ src/main/java/com/xiliulou/electricity/service/ElectricityMemberCardService.java
@@ -19,13 +19,13 @@ public interface ElectricityMemberCardService {
 
 	R queryUserList(Long offset, Long size,String productKey, String deviceName);
 
-	List<ElectricityMemberCard> queryByFranchisee(Integer id);
+	List<ElectricityMemberCard> queryByFranchisee(Long id);
 
 	R queryCount(Integer status, Integer type, Integer tenantId);
 
-	R listByFranchisee(Long offset, Long size, Integer status, Integer type, Integer tenantId, Integer franchiseeId);
+	R listByFranchisee(Long offset, Long size, Integer status, Integer type, Integer tenantId, Long franchiseeId);
 
-	R listCountByFranchisee(Integer status, Integer type, Integer tenantId, Integer franchiseeId);
+	R listCountByFranchisee(Integer status, Integer type, Integer tenantId, Long franchiseeId);
 
 	ElectricityMemberCard queryByStatus(Integer id);
 }
